refactor(services): rename file names to kebab casing and make them lowercased. Also add missing import to use-players-service hook.
